Docs support
Code to create basic documentation from the README markdown files. It
uses the current README.md structure, but should be changed so that the
README.md functionality provided by GitHub and my docs generator
shouldn't clash (ie smarter use of the standard one README.md per-folder
so that it looks good on GitHub pages and in the generated PDF).