Grab Text From Wikipedia for Test
Create the component for doing the test (not the page). Add all the
configuration and info for going out to WikiPedia and getting a random
page. Get the HTML from the page and use jsdom and jquery to get the
first few paragraphs of the article up to a configurable word count. Use
this as the test text.