Dual Provider Model Complete
The app can now load providers based on if it is in Electron or if it is running via a web page. The providers have all been written to talk to routes in an expressjs application which in turn call directly to the same configuration and repository methods that the ipc methods call.